school_api_project/requirements.txt
Django==3.2.10
djangorestframework==3.12.4
drf-spectacular==0.26.2
